unify into single file, remove SegmentSize=1 cases
