--- src/benchmarks/micro/corefx/System.Text.Json/Utf8JsonReader/Utf8JsonReader.Comments.MultiSegment.Tests.cs
@@ -1,81 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using BenchmarkDotNet.Attributes;
-using MicroBenchmarks;
-using MicroBenchmarks.Serializers;
-using System.Buffers;
-using System.Collections.Generic;
-using System.IO;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace System.Text.Json.Tests
-{
-    public partial class Utf8JsonReaderCommentsTests
-    {
-        private static ReadOnlySequence<byte> GetSequence(byte[] dataUtf8, int segmentSize)
-        {
-            int numberOfSegments = dataUtf8.Length / segmentSize + 1;
-            byte[][] buffers = new byte[numberOfSegments][];
-
-            for (int j = 0; j < numberOfSegments - 1; j++)
-            {
-                buffers[j] = new byte[segmentSize];
-                Array.Copy(dataUtf8, j * segmentSize, buffers[j], 0, segmentSize);
-            }
-
-            int remaining = dataUtf8.Length % segmentSize;
-            buffers[numberOfSegments - 1] = new byte[remaining];
-            Array.Copy(dataUtf8, dataUtf8.Length - remaining, buffers[numberOfSegments - 1], 0, remaining);
-
-            return CreateReadOnlySequence(buffers);
-        }
-
-        private static ReadOnlySequence<byte> CreateReadOnlySequence(params byte[][] buffers)
-        {
-            if (buffers.Length == 1)
-                return new ReadOnlySequence<byte>(buffers[0]);
-            var list = new List<Memory<byte>>();
-            foreach (var buffer in buffers)
-                list.Add(buffer);
-            return ReadOnlyBufferSegment.Create(list.ToArray());
-        }
-
-        private class ReadOnlyBufferSegment : ReadOnlySequenceSegment<byte>
-        {
-            public static ReadOnlySequence<byte> Create(IEnumerable<Memory<byte>> buffers)
-            {
-                ReadOnlyBufferSegment segment = null;
-                ReadOnlyBufferSegment first = null;
-                foreach (Memory<byte> buffer in buffers)
-                {
-                    var newSegment = new ReadOnlyBufferSegment()
-                    {
-                        Memory = buffer,
-                    };
-
-                    if (segment != null)
-                    {
-                        segment.Next = newSegment;
-                        newSegment.RunningIndex = segment.RunningIndex + segment.Memory.Length;
-                    }
-                    else
-                    {
-                        first = newSegment;
-                    }
-
-                    segment = newSegment;
-                }
-
-                if (first == null)
-                {
-                    first = segment = new ReadOnlyBufferSegment();
-                }
-
-                return new ReadOnlySequence<byte>(first, 0, segment, segment.Memory.Length);
-            }
-        }
-    }
-}
